Fix rename bugaboo
As Rick James said, "Cocaine is a helluva drug"; so is renaming
Emacs packages -- there be dragons lurking.  This fix allows
language major mode editing a la:

    https://orgmode.org/manual/Editing-source-code.html

to work once again.

When the user presses C-' while point is in the source block, Org
Babel is looking for the major mode `picolisp-mode` (*exactly*)
and fails to find it, absent this patch.  Why?  Because the mode
name was recently changed from picolisp-mode to plisp-mode by the
author.

Personally, I'd like to see the author just rename this package
back to picolisp-mode; short of that, this patch seems to fix the
problem with language major modes in an Org Babel document.